flask/requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
aniso8601==9.0.1
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
backcall==0.2.0
beautifulsoup4==4.12.0
bs4==0.0.1
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
chex==0.1.82
chromedriver==2.24.1
click==8.1.3
colorama==0.4.6
cssselect==1.2.0
decorator==4.4.2
distlib==0.3.6
dnspython==2.3.0
etils==1.4.1
exceptiongroup==1.1.1
executing==1.2.0
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.10.7
Flask==2.2.3
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
flatbuffers==23.5.26
flax==0.7.1
frozenlist==1.3.3
fsspec==2023.6.0
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.56.2
h11==0.14.0
h5py==3.9.0
html5lib==1.1
huggingface==0.0.1
huggingface-hub==0.16.4
idna==3.4
imageio==2.27.0
imageio-ffmpeg==0.4.8
importlib-resources==6.0.1
install==1.3.5
ipython==8.14.0
itsdangerous==2.1.2
jax==0.4.14
jaxlib==0.4.14
jedi==0.19.0
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.2.0
keras==2.13.1
libclang==16.0.6
lxml==4.9.2
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.2.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
nest-asyncio==1.5.7
networkx==3.1
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.25.2
oauthlib==3.2.2
openai==0.27.4
opt-einsum==3.3.0
optax==0.1.7
orbax-checkpoint==0.3.2
outcome==1.2.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
proglog==0.1.10
prompt-toolkit==3.0.39
protobuf==4.23.4
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.16.1
pymongo==4.3.3
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.1
rich==13.5.2
rsa==4.9
safetensors==0.3.1
scipy==1.11.1
selenium==4.8.3
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
SpeechRecognition==3.10.0
SQLAlchemy==2.0.8
stack-data==0.6.2
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
tensorstore==0.1.40
termcolor==2.3.0
tinysegmenter==0.3
tldextract==3.4.0
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.31.0
trio==0.22.0
trio-websocket==0.10.2
tweepy==4.13.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
virtualenv==20.21.0
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.2.3
wrapt==1.15.0
wsproto==1.2.0
yarl==1.8.2
zipp==3.16.2
